backend/requirements.txt
Flask
torch
torchvision
tensorflow
pydicom
Pillow
numpy
opencv-python
scikit-learn
SQLAlchemy
Flask-SQLAlchemy
Flask-CORS
python-dotenv
gunicorn
Werkzeug
marshmallow
flask-marshmallow
redis
celery